refactor(orm): add index file with clear re-exports
