+++ lib/container.js
@@ -1,11 +1,9 @@
 import { singularize, pluralize } from "inflected";
-import Db from "./db";
-import Association from "./orm/associations/association";
+import { Association, Db, Schema } from "./orm";
 import RouteHandler from "./route-handler";
 import BaseRouteHandler from "./route-handlers/base";
 import Serializer from "./serializer";
 import SerializerRegistry from "./serializer-registry";
-import Schema from "./orm/schema";
 
 const classes = {
   Db,
