+++ lib/index.js
@@ -1,25 +1,21 @@
 import Factory from "./factory";
-import IdentityManager from "./identity-manager";
+import { IdentityManager, Model, Collection, HasMany, BelongsTo } from "./orm";
 import association from "./association";
 import trait from "./trait";
 import Response from "./response";
 import Server, { createServer } from "./server";
-import Model from "./orm/model";
-import Collection from "./orm/collection";
 import Serializer from "./serializer";
 import ActiveModelSerializer from "./serializers/active-model-serializer";
 import JSONAPISerializer from "./serializers/json-api-serializer";
 import RestSerializer from "./serializers/rest-serializer";
-import HasMany from "./orm/associations/has-many";
-import BelongsTo from "./orm/associations/belongs-to";
 
 /*
   These are solely for ember-cli-mirage, a "privileged consumer", and should be
   removed once those import paths are dropped.
 */
 import _assert from "./assert";
-import _DbCollection from "./db-collection";
-import _Db from "./db";
+import _DbCollection from "./orm/db/db-collection";
+import _Db from "./orm/db/db";
 import _RouteHandler from "./route-handler";
 import _SerializerRegistry from "./serializer-registry";
 import _ormAssociationsAssociation from "./orm/associations/association";
