+++ lib/orm/index.js
@@ -1,0 +1,15 @@
+// Core ORM components
+export { default as Schema } from "./schema";
+export { default as Model } from "./model";
+export { default as Collection } from "./collection";
+export { default as PolymorphicCollection } from "./polymorphic-collection";
+
+// DB components
+export { default as Db } from "./db/db";
+export { default as DbCollection } from "./db/db-collection";
+export { default as IdentityManager } from "./db/identity-manager";
+
+// Association types
+export { default as Association } from "./associations/association";
+export { default as BelongsTo } from "./associations/belongs-to";
+export { default as HasMany } from "./associations/has-many";
