+++ lib/server.js
@@ -3,7 +3,7 @@
 import { camelize } from "./utils/inflector";
 import isAssociation from "./utils/is-association";
 import assert from "./assert";
-import BelongsTo from "./orm/associations/belongs-to";
+import { BelongsTo } from "./orm";
 import Container from "./container";
 import { singularize, pluralize } from "inflected";
 import pick from "lodash/pick.js";
